function_app/pyproject.toml
[project]
name = "Copilot-Plugins-Azure-Functions-deployment-v2"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "Sumsam Ali", email = "sumsam.ali@imperiumdynamics.com"},
]
dependencies = [
    "azure-functions",
    "requests",
    "azure-identity",
    "python-dateutil",
    "pytz",
    "validate-email-address",
    "fuzzywuzzy",
    "aiohttp>=3.13.2",
    "pydantic[email]>=2.12.4",
    "pydantic-settings>=2.12.0",
]
requires-python = "==3.11.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false